requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
awscli==1.27.69
awsebcli==3.19.0
bcrypt==4.0.1
binaryornot==0.4.4
blessed==1.20.0
boto3==1.26.66
botocore==1.17.63
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.7
chevron==0.14.0
click==8.0.3
colorama==0.4.6
cookiecutter==2.1.1
cryptography==39.0.0
dateparser==1.1.5
distro==1.8.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
docker==3.7.3
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.1.0
Flask==2.0.2
frozenlist==1.3.3
future==0.16.0
idna==2.6
itsdangerous==2.0.1
Jinja2==3.0.2
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==2.6.0
MarkupSafe==2.0.1
multidict==6.0.4
numpy==1.21.3
openai==0.26.5
openpyxl==3.0.9
packaging==23.0
pandas==1.3.4
pandas-stubs==1.2.0.35
paramiko==3.0.0
pathspec==0.5.9
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyrsistent==0.19.3
python-dateutil==2.8.0
python-dotenv==0.19.2
python-slugify==7.0.0
pytz==2021.3
PyYAML==5.3.1
regex==2021.9.30
requests==2.20.1
rsa==3.1.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
semantic-version==2.5.0
serverlessrepo==0.1.10
six==1.11.0
sqlparse==0.4.3
tabulate==0.7.5
termcolor==1.1.0
text-unidecode==1.3
texttable==0.9.1
toml==0.10.2
tomlkit==0.7.2
tqdm==4.62.3
tzlocal==3.0
urllib3==1.24.3
watchdog==2.1.2
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.2
yarl==1.8.2
